Fixed cut, paste and copy for open shadow dom
